models.py: Move DynamicSiteMiddleware methods to AliasManager:
* DynamicSiteMiddleware.get_site() -> AliasManager.resolve()
    AliasManager.resolve now returns an Alias object, not a Site
    object.

* DynamicSiteMiddleware._expand_netloc() -> AliasManager._expand_netloc()
    Implementation unchanged.
